test: try to remove repeats
